move searchView function to detail league activity
