+++ app/src/main/java/com/dicoding/picodiploma/sejiwaproject/features/league/detail/DetailView.kt
@@ -3,7 +3,5 @@ package com.dicoding.picodiploma.sejiwaproject.features.league.detail
 import com.dicoding.picodiploma.sejiwaproject.features.league.detail.model.League
 
 interface DetailView {
-    fun showLoading()
-    fun hideLoading()
     fun showDetailList(data: List<League>)
 }
