+++ app/src/main/java/com/dicoding/picodiploma/sejiwaproject/features/league/list/LeagueListActivity.kt
@@ -1,14 +1,10 @@
 package com.dicoding.picodiploma.sejiwaproject.features.league.list
 
-import android.content.Intent
 import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.dicoding.picodiploma.sejiwaproject.R
-import com.dicoding.picodiploma.sejiwaproject.features.match.search.SearchMatchActivity
 
 class LeagueListActivity : AppCompatActivity() {
     private lateinit var rvLeague: RecyclerView
@@ -50,21 +46,6 @@ class LeagueListActivity : AppCompatActivity() {
             )
         }
         logo.recycle()
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.options_menu, menu)
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when(item.itemId) {
-            R.id.search -> {
-                val moveIntent = Intent(this@LeagueListActivity, SearchMatchActivity::class.java)
-                startActivity(moveIntent)
-            }
-        }
-        return super.onOptionsItemSelected(item)
     }
 }
